#135 Осуществления функционала файлообмена в чате и внедрения эмоджи - Создал модель файлов - Упорядочил код представлений чата
