+++ chat/admin.py
@@ -1,6 +1,8 @@
 from django.contrib import admin
-from .models import ChatRoom, ChatRoomMembership, ChatMessage
+
+from chat.models import ChatRoom, ChatRoomMembership, ChatMessage, File
 
 admin.site.register(ChatRoom)
 admin.site.register(ChatRoomMembership)
 admin.site.register(ChatMessage)
+admin.site.register(File)
