+++ chat/views/__init__.py
@@ -1,0 +1,3 @@
+from chat.views.centrifugo import connect, publish, subscribe, upload_file
+from chat.views.group import GroupDetailView, CreateGroupChatView, UpdateGroupChatView
+from chat.views.private import RoomView, StartChatView, ChatsView
